Update pal from commit: e88a0e82
* RGP Profiling Doesn't Work on RX 6800
* Fix a bug in InterfaceLogger related to IndirectCmdGenerator
* Set DefaultSqttDetailedCuIndex scope to PrivatePalKey
* Compartmentalizing Pm4 Cleanup
* Fix bug in ArchiveFile for PushBack
* Relax typed buffer stride requirement
* Debug logging support in PAL
* Couple of barrier log and acquire release barrier changes
* modify Accum factor settings to be reread
* Performance optimization by setting binning panel keys
* GpuProfiler Lifetime bug in CTV/DSV
* Fix incorrect logic of optimizing gfxWriteCachesDirty flag for CmdRelease/CmdAcquire path
* Move most of the GfxCmdBufferState to Pm4CmdBuffer and rename to Pm4CmdBufferState
* Fix ABI Refactor Regression
* Initialize RMT pipeline flags
* Add and report additional device properties, bump RGP ASIC_INFO version
* Fix code object database missing entries
* Compartmentalize PM4/PUP/SDMA (GfxCmdBuf)
* Pipeline ABI Refactor, Part 1 (PAL - Gfx9 - Graphics only)
* Expose vgtMultiPrimIbResetEn MATCH_ALL_BITS
* Report the HW LDS granularity allocation back to the client
* Implement 'TossPointAfterRaster' setting
* Fix HashString to treat bytes as unsigned
* Performance optimization by setting binning panel keys
* Util::CheckSequential causes weird intellisense static_assert bug
* Revert an ACE wait-for-idle optimization
* Improve barrier log format
* Acquire release barrier code cleanup and improvement
* update devdriver to v22.06.17
* Add and report MALL cache size
* Fix miss VS stall for IB->CsUav transition in CmdBarrier path
* acquire_mem/release_mem packet builder refactor
* Full Any-Asic-IFH for Linux
* Add resource update token logging support for RMV
* Fix Vulkan SwissKnife failure on Vega56 due to enable acquire release barrier

+++ inc/util/palInlineFuncs.h
@@ -915,7 +915,7 @@ constexpr uint32 HashString(
 
     for (uint32 i = 0; i < strSize; i++)
     {
-        hash ^= static_cast<uint32>(pStr[i]);
+        hash ^= uint8(pStr[i]);
         hash *= FnvPrime;
     }
